[INERTNAL] Added a lot of documentation and changed the header file structure.
